Temporary fix to test script error
The test script uses the 'react-scripts' package, which contains utilites
for Create React App. This is known to have a whole host of problems,
which is why it was removed in other branches, and we use Vite instead.
However, in order to patch this fix for now (before we merge the correct
branches and edit this file again), I have set the CI env variable, and
changed the `--watch` option to `--watchAll`. Hopefully this fixes the
issue and means that we won't use up all our GitHub Action minutes.