Create initial Object-Relational Mapping (#13)
This PR sets up the database structure and ORM used to interact with it, as well as systems to interact with AWS S3 and STS.

Using the AWS connectors and functions relies on having an AWS configuration available. This means providing an AWS_ACCESS_KEY_ID a AWS_SECRET_ACCESS_KEY, and an AWS_REGION via Environment variables or the AWS CLI.
Should anyone want to use these functions, talk to me (Enzo Bestetti) and I will provide the currently available credentials.

Co-authored-by: Markus Meresma <markus.meresma@gmail.com>