Add File storage functionality.
In this commit I introduce a wrapper for file storage to be used throughout the application.
It provides an abstraction from whichever storage service we use, including whether we are storing files locally or remotely.
From now on, you can call the functions from StorageService to store, get, and delete files.