Introduce Regexes to "unescape" quotations.
Inconsistency in whether quotations are escaped in LLM responses lead to undefined behaviour. The program may work, or exceptions may be thrown. We aim to control that by ensuring no quotations are escaped.

Additionally, the C-style comment Regex has been updated to use negative look-behind to ensure we are not accidentally matching a URL.

Python-style comment cleaning has also been removed in this commit because it is not being used at this time.